Homogenize all REST calls being made to CSE server from client
