+++ container_service_extension/client/cse_client/api_33/native_cluster_api.py
@@ -5,8 +5,6 @@
 import pyvcloud.vcd.client as vcd_client
 
 from container_service_extension.client.cse_client.cse_client import CseClient
-from container_service_extension.client.request_maker import make_request
-from container_service_extension.client.response_processor import process_response  # noqa: E501
 import container_service_extension.common.constants.shared_constants as shared_constants  # noqa: E501
 
 
@@ -34,37 +32,34 @@ class NativeClusterApi(CseClient):
     def list_clusters(self, filters=None):
         if filters is None:
             filters = {}
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=self._clusters_uri,
             method=shared_constants.RequestMethod.GET,
             params=filters,
             accept_type='application/json')
-        return process_response(response)
+        return self.process_response(response)
 
     def get_cluster(self, cluster_name, filters=None):
         if filters is None:
             filters = {}
         uri = f'{self._cluster_uri}/{cluster_name}'
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.GET,
             params=filters,
             accept_type='application/json')
-        return process_response(response)
+        return self.process_response(response)
 
     def get_cluster_upgrade_plan(self, cluster_name, filters=None):
         if filters is None:
             filters = {}
         uri = f'{self._cluster_uri}/{cluster_name}/upgrade-plan'
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.GET,
             params=filters,
             accept_type='application/json')
-        return process_response(response)
+        return self.process_response(response)
 
     def upgrade_cluster(self, cluster_name, template_name, template_revision,
                         org_name=None, ovdc_name=None):
@@ -76,14 +71,13 @@ class NativeClusterApi(CseClient):
             shared_constants.RequestKey.ORG_NAME: org_name,
             shared_constants.RequestKey.OVDC_NAME: ovdc_name,
         }
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.POST,
             accept_type='application/json',
             media_type='application/json',
             payload=payload)
-        return process_response(response)
+        return self.process_response(response)
 
     def create_cluster(self, cluster_name, ovdc_name, network_name, node_count=None,  # noqa: E501
                        org_name=None, cpu=None, memory=None,
@@ -104,14 +98,13 @@ class NativeClusterApi(CseClient):
             shared_constants.RequestKey.ROLLBACK: rollback,
             shared_constants.RequestKey.ORG_NAME: org_name
         }
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=self._clusters_uri,
             method=shared_constants.RequestMethod.POST,
             accept_type='application/json',
             media_type='application/json',
             payload=payload)
-        return process_response(response)
+        return self.process_response(response)
 
     def update_cluster(self, cluster_name, network_name, node_count,
                        org_name=None, ovdc_name=None, template_name=None,
@@ -131,38 +124,35 @@ class NativeClusterApi(CseClient):
             shared_constants.RequestKey.MB_MEMORY: memory,
             shared_constants.RequestKey.SSH_KEY: ssh_key
         }
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.PUT,
             accept_type='application/json',
             media_type='application/json',
             payload=payload)
-        return process_response(response)
+        return self.process_response(response)
 
     def delete_cluster(self, cluster_name, filters=None):
         if filters is None:
             filters = {}
         uri = f"{self._cluster_uri}/{cluster_name}"
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.DELETE,
             params=filters,
             accept_type='application/json')
-        return process_response(response)
+        return self.process_response(response)
 
     def get_cluster_config(self, cluster_name, filters=None):
         if filters is None:
             filters = {}
         uri = f"{self._cluster_uri}/{cluster_name}/config"
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.GET,
             params=filters,
             accept_type='application/json')
-        return process_response(response)
+        return self.process_response(response)
 
     def add_node(self, cluster_name, network_name, node_count=1, org_name=None,
                  ovdc_name=None, cpu=None, memory=None, storage_profile=None,
@@ -183,27 +173,25 @@ class NativeClusterApi(CseClient):
             shared_constants.RequestKey.ENABLE_NFS: enable_nfs,
             shared_constants.RequestKey.ROLLBACK: rollback
         }
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=self._nodes_uri,
             method=shared_constants.RequestMethod.POST,
             accept_type='application/json',
             media_type='application/json',
             payload=payload)
-        return process_response(response)
+        return self.process_response(response)
 
     def get_node_info(self, cluster_name, node_name, filters=None):
         if filters is None:
             filters = {}
         uri = f"{self._node_uri}/{node_name}"
         filters[shared_constants.RequestKey.CLUSTER_NAME] = cluster_name
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.GET,
             params=filters,
             accept_type='application/json')
-        return process_response(response)
+        return self.process_response(response)
 
     def delete_nodes(self, cluster_name, nodes_list,
                      org_name=None, ovdc_name=None):
@@ -213,11 +201,10 @@ class NativeClusterApi(CseClient):
             shared_constants.RequestKey.OVDC_NAME: ovdc_name,
             shared_constants.RequestKey.NODE_NAMES_LIST: nodes_list
         }
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=self._nodes_uri,
             method=shared_constants.RequestMethod.DELETE,
             accept_type='application/json',
             media_type='application/json',
             payload=payload)
-        return process_response(response)
+        return self.process_response(response)
