+++ container_service_extension/client/cse_client/api_35/native_cluster_api.py
@@ -8,8 +8,6 @@ from typing import Dict, List
 import pyvcloud.vcd.client as vcd_client
 
 from container_service_extension.client.cse_client.cse_client import CseClient
-from container_service_extension.client.request_maker import make_request
-import container_service_extension.client.response_processor as response_processor  # noqa: E501
 import container_service_extension.common.constants.shared_constants as shared_constants  # noqa: E501
 import container_service_extension.rde.models.common_models as common_models
 import container_service_extension.rde.models.rde_1_0_0 as rde_1_0_0
@@ -26,86 +24,75 @@ class NativeClusterApi(CseClient):
     def create_cluster(self, cluster_entity_definition: rde_1_0_0.NativeEntity):  # noqa: E501
         cluster_entity_dict = asdict(cluster_entity_definition)
         uri = self._clusters_uri
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.POST,
             accept_type='application/json',
             media_type='application/json',
             payload=cluster_entity_dict)
 
-        return common_models.DefEntity(
-            **response_processor.process_response(response))
+        return common_models.DefEntity(**self.process_response(response))
 
     def update_cluster_by_cluster_id(self, cluster_id, cluster_entity_definition: rde_1_0_0.NativeEntity):  # noqa: E501
         cluster_entity_dict = asdict(cluster_entity_definition)
         uri = f"{self._cluster_uri}/{cluster_id}"
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.PUT,
             accept_type='application/json',
             media_type='application/json',
             payload=cluster_entity_dict)
 
-        return common_models.DefEntity(
-            **response_processor.process_response(response))
+        return common_models.DefEntity(**self.process_response(response))
 
     def delete_cluster_by_cluster_id(self, cluster_id):
         uri = f"{self._cluster_uri}/{cluster_id}"
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.DELETE,
             accept_type='application/json')
 
-        return common_models.DefEntity(**response_processor.process_response(response))  # noqa: E501
+        return common_models.DefEntity(**self.process_response(response))
 
     def delete_nfs_node_by_node_name(self, cluster_id: str, node_name: str):
         uri = f"{self._cluster_uri}/{cluster_id}/nfs/{node_name}"
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.DELETE,
             accept_type='application/json')
 
-        return common_models.DefEntity(
-            **response_processor.process_response(response))
+        return common_models.DefEntity(**self.process_response(response))
 
     def get_cluster_config_by_cluster_id(self, cluster_id: str) -> dict:
         uri = f"{self._cluster_uri}/{cluster_id}/config"
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.GET,
             accept_type='application/json')
 
-        return response_processor.process_response(response)
+        return self.process_response(response)
 
     def get_upgrade_plan_by_cluster_id(self, cluster_id: str):
         uri = f'{self._cluster_uri}/{cluster_id}/upgrade-plan'
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.GET,
             accept_type='application/json')
 
-        return response_processor.process_response(response)
+        return self.process_response(response)
 
     def upgrade_cluster_by_cluster_id(self, cluster_id: str,
                                       cluster_upgrade_definition: common_models.DefEntity):  # noqa: E501
         uri = f'{self._uri}/cluster/{cluster_id}/action/upgrade'
         entity_dict = asdict(cluster_upgrade_definition.entity)
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.POST,
             accept_type='application/json',
             media_type='application/json',
             payload=entity_dict)
 
-        return common_models.DefEntity(
-            **response_processor.process_response(response))
+        return common_models.DefEntity(**self.process_response(response))
 
     def get_single_page_cluster_acl(self, cluster_id,
                                     page=shared_constants.CSE_PAGINATION_FIRST_PAGE_NUMBER,  # noqa: E501
@@ -115,15 +102,13 @@ class NativeClusterApi(CseClient):
             shared_constants.PaginationKey.PAGE_NUMBER: page,
             shared_constants.PaginationKey.PAGE_SIZE: page_size
         }
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=query_uri,
             method=shared_constants.RequestMethod.GET,
             params=filters,
             accept_type='application/json')
 
-        processed_response = response_processor.process_response(response)
-        return processed_response
+        return self.process_response(response)
 
     def list_native_cluster_acl_entries(self, cluster_id):
         page_num = 0
@@ -144,12 +129,11 @@ class NativeClusterApi(CseClient):
         payload = {
             shared_constants.ClusterAclKey.ACCESS_SETTING: acl_entries
         }
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.PUT,
             accept_type='application/json',
             media_type='application/json',
             payload=payload)
 
-        response_processor.process_response(response)
+        self.process_response(response)
