+++ container_service_extension/client/cse_client/api_35/ovdc_api.py
@@ -7,8 +7,6 @@ from dataclasses import asdict
 import pyvcloud.vcd.client as vcd_client
 
 from container_service_extension.client.cse_client.cse_client import CseClient
-from container_service_extension.client.request_maker import make_request
-from container_service_extension.client.response_processor import process_response  # noqa: E501
 import container_service_extension.common.constants.shared_constants as shared_constants  # noqa: E501
 import container_service_extension.rde.models.common_models as common_models
 
@@ -20,8 +18,9 @@ class OvdcApi(CseClient):
                     f"{shared_constants.CSE_3_0_URL_FRAGMENT}"
         self._org_vdcs_uri = f"{self._uri}/orgvdcs"
         self._ovdc_uri = f"{self._uri}/ovdc"
-        # NOTE: The request page size is overrided because the CSE server takes
-        # an average of 10 seconds (Default vCD timeout) if there are 5 OVDCs
+        # NOTE: The request page size is overridden because the CSE server
+        # takes an average of 10 seconds (Default vCD timeout) if there are
+        # 5 OVDCs
         self._request_page_size = 10
 
     def get_all_ovdcs(self):
@@ -37,22 +36,20 @@ class OvdcApi(CseClient):
 
     def get_ovdc(self, ovdc_id):
         uri = f"{self._ovdc_uri}/{ovdc_id}"
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.GET,
             accept_type='application/json')
 
-        return common_models.Ovdc(**process_response(response))
+        return common_models.Ovdc(**self.process_response(response))
 
     def update_ovdc(self, ovdc_id, ovdc_obj: common_models.Ovdc):
         uri = f"{self._ovdc_uri}/{ovdc_id}"
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.PUT,
             accept_type='application/json',
             media_type='application/json',
             payload=asdict(ovdc_obj))
 
-        return process_response(response)
+        return self.process_response(response)
