+++ container_service_extension/client/cse_client/pks_cluster_api.py
@@ -5,8 +5,6 @@
 import pyvcloud.vcd.client as vcd_client
 
 from container_service_extension.client.cse_client.cse_client import CseClient
-from container_service_extension.client.request_maker import make_request
-import container_service_extension.client.response_processor as response_processor  # noqa: E501
 import container_service_extension.common.constants.shared_constants as shared_constants  # noqa: E501
 
 
@@ -20,25 +18,23 @@ class PksClusterApi(CseClient):
     def list_clusters(self, filters=None):
         if filters is None:
             filters = {}
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=self._clusters_uri,
             method=shared_constants.RequestMethod.GET,
             params=filters,
             accept_type='application/json')
-        return response_processor.process_response(response)
+        return self.process_response(response)
 
     def get_cluster(self, cluster_name, filters=None):
         if filters is None:
             filters = {}
         uri = f'{self._cluster_uri}/{cluster_name}'
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.GET,
             params=filters,
             accept_type='application/json')
-        return response_processor.process_response(response)
+        return self.process_response(response)
 
     def create_cluster(self, cluster_name, ovdc_name, node_count=None, org_name=None):  # noqa: E501
         payload = {
@@ -47,15 +43,13 @@ class PksClusterApi(CseClient):
             shared_constants.RequestKey.OVDC_NAME: ovdc_name,
             shared_constants.RequestKey.ORG_NAME: org_name
         }
-        uri = self._clusters_uri
-        response = make_request(
-            client=self._client,
-            uri=uri,
+        response = self.do_request(
+            uri=self._clusters_uri,
             method=shared_constants.RequestMethod.POST,
             accept_type='application/json',
             media_type='application/json',
             payload=payload)
-        return response_processor.process_response(response)
+        return self.process_response(response)
 
     def update_cluster(self, cluster_name, node_count, org_name=None,
                        ovdc_name=None):
@@ -66,29 +60,26 @@ class PksClusterApi(CseClient):
             shared_constants.RequestKey.OVDC_NAME: ovdc_name
         }
         uri = f"{self._cluster_uri}/{cluster_name}"
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.PUT,
             accept_type='application/json',
             media_type='application/json',
             payload=payload)
-        return response_processor.process_response(response)
+        return self.process_response(response)
 
     def delete_cluster(self, cluster_name, org_name=None, ovdc_name=None):
         uri = f"{self._cluster_uri}/{cluster_name}"
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.DELETE,
             accept_type='application/json')
-        return response_processor.process_response(response)
+        return self.process_response(response)
 
     def get_cluster_config(self, cluster_name, org_name=None, ovdc_name=None):
         uri = f"{self._cluster_uri}/{cluster_name}/config"
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=uri,
             method=shared_constants.RequestMethod.GET,
             accept_type='application/json')
-        return response_processor.process_response(response)
+        return self.process_response(response)
