+++ container_service_extension/client/cse_client/templates_api.py
@@ -5,8 +5,6 @@
 import pyvcloud.vcd.client as vcd_client
 
 from container_service_extension.client.cse_client.cse_client import CseClient
-from container_service_extension.client.request_maker import make_request
-from container_service_extension.client.response_processor import process_response  # noqa: E501
 import container_service_extension.common.constants.shared_constants as shared_constants  # noqa: E501
 
 
@@ -17,9 +15,8 @@ class TemplatesApi(CseClient):
         self._templates_uri = f"{self._uri}/templates"
 
     def list_templates(self):
-        response = make_request(
-            client=self._client,
+        response = self.do_request(
             uri=self._templates_uri,
             method=shared_constants.RequestMethod.GET,
             accept_type='application/json')
-        return process_response(response)
+        return self.process_response(response)
